pyproject.toml
[project]
name = "celine-policies"
version = "0.1.0"
description = "CELINE Policy Authorization Service with embedded OPA"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "pyjwt[crypto]>=2.10.0",
    "httpx>=0.28.0",
    "cachetools>=5.5.0",
    "structlog>=25.5.0",
    "celine-regorus>=0.5.0.post20260129052409",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
]


[tool.uv]  # optional, helps uv handle locking
package = true

[tool.setuptools.packages.find]
where = ["src"]
include = ["celine.policies*"]
exclude = ["tests*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]

[dependency-groups]
dev = [
    "dumpster-cli>=1.4.0",
    "pytest-asyncio>=1.3.0",
]


[tool.semantic_release]
version_toml = ["pyproject.toml:project.version"]
update_pyproject_toml = true 
commit_parser = "conventional"

build_command = """
    uv lock --upgrade-package "$PACKAGE_NAME"
    git add uv.lock
    uv build
"""

[tool.semantic_release.commit_parser_options]
minor_tags = ["feat"]
patch_tags = ["fix", "perf", "chore"]
parse_squash_commits = true
ignore_merge_commits = true

